Move tag and language parsing to selectors
