+++ client/src/views/all-program-items/allProgramItemsSlice.ts
@@ -1,8 +1,11 @@
-import { createSlice, PayloadAction } from "@reduxjs/toolkit";
+import { createSelector, createSlice, PayloadAction } from "@reduxjs/toolkit";
+import { uniq } from "lodash-es";
 import { AllProgramItemsState, RootState } from "client/types/reduxTypes";
 import {
+  Language,
   ProgramItem,
   ProgramItemWithUserSignups,
+  Tag,
 } from "shared/types/models/programItem";
 
 const initialState: AllProgramItemsState = {
@@ -42,3 +45,48 @@ export const allProgramItemsReducer = allProgramItemsSlice.reducer;
 
 export const selectProgramItems = (state: RootState): readonly ProgramItem[] =>
   state.allProgramItems.programItems;
+
+export const selectTags = createSelector(
+  [selectProgramItems],
+  (programItems) => {
+    const ignoredTags = [Tag.GUEST_OF_HONOR, Tag.THEME_MONSTERS];
+
+    const tags = uniq([
+      Tag.BEGINNER_FRIENDLY,
+      Tag.ALL_AGES,
+      Tag.AIMED_UNDER_13,
+      Tag.AIMED_BETWEEN_13_17,
+      Tag.AIMED_ADULTS,
+      Tag.FOR_18_PLUS_ONLY,
+      ...Object.values(Tag),
+    ]);
+    return tags
+      .filter((tag) => {
+        return programItems.some((programItem) =>
+          programItem.tags.includes(tag),
+        );
+      })
+      .filter((tag) => !ignoredTags.includes(tag));
+  },
+);
+
+export const selectLanguages = createSelector(
+  [selectProgramItems],
+  (programItems) => {
+    const ignoredLanguages = [Language.LANGUAGE_FREE];
+
+    const languages = uniq([
+      Language.FINNISH,
+      Language.ENGLISH,
+      Language.SWEDISH,
+      ...Object.values(Language),
+    ]);
+    return languages
+      .filter((language) => {
+        return programItems.some((programItem) =>
+          programItem.languages.includes(language),
+        );
+      })
+      .filter((language) => !ignoredLanguages.includes(language));
+  },
+);
